requirements.txt
Django==4.1.5
django-crispy-forms==1.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
pre-commit==3.0.0